requirements.txt
boto3
snowflake-snowpark-python
python-dotenv